code tidying and restructuring
