+++ views/index.hbs
@@ -54,6 +54,7 @@
 
 
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="/javascripts/words.js"></script>
 <script src="/javascripts/word_invade.js"></script>
